+++ src/main/java/net/frozenblock/wilderwild/datagen/advancement/WWAdvancementProvider.java
@@ -18,28 +18,42 @@
 
 package net.frozenblock.wilderwild.datagen.advancement;
 
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricAdvancementProvider;
 import net.frozenblock.wilderwild.WWConstants;
 import net.frozenblock.wilderwild.advancement.FragileIceFallOntoAndBreakTrigger;
+import net.frozenblock.wilderwild.advancement.GeyserPushMobTrigger;
 import net.frozenblock.wilderwild.advancement.MobBottleTrigger;
 import net.frozenblock.wilderwild.advancement.TermiteEatTrigger;
+import net.frozenblock.wilderwild.block.state.properties.GeyserType;
 import net.frozenblock.wilderwild.registry.WWBlocks;
+import net.frozenblock.wilderwild.registry.WWEntityTypes;
 import net.frozenblock.wilderwild.registry.WWItems;
+import net.frozenblock.wilderwild.tag.WWEntityTags;
 import net.minecraft.advancements.Advancement;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementType;
 import net.minecraft.advancements.critereon.BlockPredicate;
+import net.minecraft.advancements.critereon.EntityFlagsPredicate;
+import net.minecraft.advancements.critereon.EntityPredicate;
+import net.minecraft.advancements.critereon.EntityTypePredicate;
 import net.minecraft.advancements.critereon.FilledBucketTrigger;
 import net.minecraft.advancements.critereon.InventoryChangeTrigger;
 import net.minecraft.advancements.critereon.ItemPredicate;
+import net.minecraft.advancements.critereon.MinMaxBounds;
+import net.minecraft.advancements.critereon.MovementPredicate;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.HolderSet;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.world.entity.EntityType;
 
 public final class WWAdvancementProvider extends FabricAdvancementProvider {
+
 	public WWAdvancementProvider(FabricDataOutput output, CompletableFuture<HolderLookup.Provider> registries) {
 		super(output, registries);
 	}
@@ -48,6 +62,8 @@ public final class WWAdvancementProvider extends FabricAdvancementProvider {
 	public void generateAdvancement(HolderLookup.Provider registries, Consumer<AdvancementHolder> writer) {
 		AdvancementHolder adventure = Advancement.Builder.advancement().build(WWConstants.vanillaId("adventure/root"));
 		AdvancementHolder husbandry = Advancement.Builder.advancement().build(WWConstants.vanillaId("husbandry/root"));
+
+		HolderLookup.RegistryLookup<EntityType<?>> entityTypeRegistryLookup = registries.lookupOrThrow(Registries.ENTITY_TYPE);
 
 		Advancement.Builder.advancement()
 			.parent(husbandry)
@@ -123,6 +139,81 @@ public final class WWAdvancementProvider extends FabricAdvancementProvider {
 			)
 			.addCriterion("termite_ate_block", TermiteEatTrigger.TriggerInstance.termiteEat(BlockPredicate.Builder.block().of(BlockTags.OVERWORLD_NATURAL_LOGS), true))
 			.save(writer, WWConstants.string("adventure/use_termite_on_tree"));
+
+		AdvancementHolder geyserPushedFlightlessBird = Advancement.Builder.advancement()
+			.parent(adventure)
+			.display(
+				WWBlocks.GEYSER,
+				Component.translatable("wilderwild.advancements.adventure.geyser_pushed_flightless_bird.title"),
+				Component.translatable("wilderwild.advancements.adventure.geyser_pushed_flightless_bird.description"),
+				null,
+				AdvancementType.TASK,
+				true,
+				true,
+				false
+			)
+			.addCriterion("geyser_pushed_mob", GeyserPushMobTrigger.TriggerInstance.geyserPushMob(
+				Optional.of(
+					EntityPredicate.Builder.entity()
+						.entityType(
+							new EntityTypePredicate(
+								HolderSet.direct(
+									EntityType.CHICKEN.builtInRegistryHolder(),
+									WWEntityTypes.OSTRICH.builtInRegistryHolder()
+								)
+							)
+						)
+						.moving(
+							new MovementPredicate(
+								MinMaxBounds.Doubles.ANY,
+								MinMaxBounds.Doubles.atLeast(0.5D),
+								MinMaxBounds.Doubles.ANY,
+								MinMaxBounds.Doubles.ANY,
+								MinMaxBounds.Doubles.ANY,
+								MinMaxBounds.Doubles.ANY,
+								MinMaxBounds.Doubles.ANY
+							)
+						)
+						.build()
+				),
+				true,
+				GeyserType.AIR
+				)
+			)
+			.save(writer, WWConstants.string("adventure/geyser_pushed_flightless_bird"));
+
+		Advancement.Builder.advancement()
+			.parent(geyserPushedFlightlessBird)
+			.display(
+				WWBlocks.GEYSER,
+				Component.translatable("wilderwild.advancements.adventure.geyser_sets_cow_on_fire.title"),
+				Component.translatable("wilderwild.advancements.adventure.geyser_sets_cow_on_fire.description"),
+				null,
+				AdvancementType.TASK,
+				true,
+				true,
+				false
+			)
+			.addCriterion("geyser_pushed_mob", GeyserPushMobTrigger.TriggerInstance.geyserPushMob(
+				Optional.of(
+					EntityPredicate.Builder.entity()
+						.entityType(
+							new EntityTypePredicate(
+								HolderSet.direct(
+									EntityType.COW.builtInRegistryHolder(),
+									EntityType.MOOSHROOM.builtInRegistryHolder(),
+									WWEntityTypes.MOOBLOOM.builtInRegistryHolder()
+								)
+							)
+						)
+						.flags(EntityFlagsPredicate.Builder.flags().setOnFire(true))
+						.build()
+				),
+				true,
+				GeyserType.LAVA
+				)
+			)
+			.save(writer, WWConstants.string("adventure/geyser_sets_cow_on_fire"));
 
 		Advancement.Builder.advancement()
 			.parent(Advancement.Builder.advancement().build(WWConstants.vanillaId("adventure/walk_on_powder_snow_with_leather_boots")))